chore(router): Move useMatch to its own file
