+++ packages/router/src/index.ts
@@ -20,6 +20,7 @@ export { default as RouteFocus } from './route-focus'
 export * from './route-focus'
 export * from './useRouteName'
 export * from './useRoutePaths'
+export * from './useMatch'
 
 export { parseSearch, getRouteRegexAndParams, matchPath } from './util'
